requirements.txt
# Vector DB + OpenAI
chromadb>=0.4.22
openai>=1.12.0
python-dotenv>=1.0.0

# GitHub klonlama
GitPython>=3.1.40

# RAG / parçalama / vektör store köprüsü
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-google-genai>=2.1.0
langchain-chroma>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# UI ve çıktı (proje tanımına göre)
streamlit>=1.30.0
python-docx>=1.1.0
pypdf>=4.0.0
